Update makefile, and package path
